requirements.txt
azure-storage-file-datalake
azure-storage-blob
